% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcFlu20192020-package.R
\docType{package}
\name{cdcFlu20192020}
\alias{cdcFlu20192020}
\alias{cdcFlu20192020-package}
\title{cdcFlu20192020 - Code and submissions for 2019-2020 CDC flu prediction contest}
\description{
Code and submissions for 2019-2020 CDC flu prediction contest
}
\author{
Evan Ray <eray@mtholyoke.edu>
}
