% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcfluforecasts-package.R
\docType{package}
\name{cdcfluforecasts}
\alias{cdcfluforecasts}
\alias{cdcfluforecasts-package}
\title{cdcfluforecasts - Code and submissions for CDC FluSight prediction contest from ReichLab/Kernel of Truth team}
\description{
Code and submissions for CDC FluSight prediction contest
}
\author{
Evan Ray <eray@mtholyoke.edu>, Nicholas Reich <nick@umass.edu>
}
